feat: 획득 경험치 입력 구현
