+++ be15-4th-b612-littleprince-be/src/main/java/littleprince/member/command/mapper/ExpHistoryCommandMapper.java
@@ -8,5 +8,5 @@ public interface ExpHistoryCommandMapper {
     void insertExpHistory(@Param("memberId") Long memberId, @Param("expPoint") int expPoint);
 
 
-    void insertExpHistoryId(Long memberId, int amount, Long projectId);
+    void insertExpHistoryId(@Param("memberId") Long memberId, @Param("expPoint") int expPoint, Long projectId);
 }
