+++ be15-4th-b612-littleprince-be/src/main/java/littleprince/plan/command/domain/aggregate/Task.java
@@ -32,19 +32,6 @@ public class Task {
 
     private LocalDate date;
 
-    @Override
-    public String toString() {
-        return "Task{" +
-                "taskId=" + taskId +
-                ", memberId=" + memberId +
-                ", projectId=" + projectId +
-                ", content='" + content + '\'' +
-                ", checked=" + checked +
-                ", date=" + date +
-                ", createdAt=" + createdAt +
-                '}';
-    }
-
     private LocalDateTime createdAt;
 
     @Builder
