+++ be15-4th-b612-littleprince-be/src/main/resources/mappers/member/ExpHistoryQueryMapper.xml
@@ -18,7 +18,6 @@
             count(*) totalExp
           FROM task
         WHERE member_id = #{ memberId }
-          AND project_id IS NULL
           AND is_checked = 'Y'
           AND date = CURDATE() - INTERVAL 1 DAY
     </select>
